--- a/AUTHORS
+++ b/AUTHORS
@@ -30,6 +30,7 @@ David Stainton <dstainton415@gmail.com>
 Jesse Ward <jesse@jesseward.com>
 Kane Mathers <kane@kanemathers.name>
 Jose Selvi <jselvi@pentester.es>
+Yerden Zhumabekov <yerden.zhumabekov@gmail.com>
 
 -----------------------------------------------
 FORKED FROM github.com/akrennmair/gopcap