`hdr` struct instance in `pcapOfflineFilter()` escapes through Cgo call. Preclude pointer escaping by placing `pcapPkthdr` inside `BPF` struct, thus allocating it on the heap and by performing pointer conversions.

Also added a benchmark to test this.

```
benchmark                             old ns/op     new ns/op     delta
BenchmarkPcapNonmatchingBPFFilter     999           756           -24.32%
BenchmarkPcapMatchingBPFFilter        1113          884           -20.58%

benchmark                             old allocs     new allocs     delta
BenchmarkPcapNonmatchingBPFFilter     1              0              -100.00%
BenchmarkPcapMatchingBPFFilter        1              0              -100.00%

benchmark                             old bytes     new bytes     delta
BenchmarkPcapNonmatchingBPFFilter     32            0             -100.00%
BenchmarkPcapMatchingBPFFilter        32            0             -100.00%
```